requirements.txt
urllib3
ujson
uuid
boto3
kafka-python
pyopenssl
asyncio
aiosfstream